requirements.txt
langchain
openai
python-dotenv
langchain-google-genai
langchain_community
mypy